+++ src/main/java/org/opensearch/knn/plugin/transport/ClearCacheAction.java
@@ -6,7 +6,7 @@
 package org.opensearch.knn.plugin.transport;
 
 import org.opensearch.action.ActionType;
-import org.opensearch.common.io.stream.Writeable;
+import org.opensearch.core.common.io.stream.Writeable;
 
 /**
  * Action associated with ClearCache