+++ src/main/java/org/opensearch/knn/plugin/transport/ClearCacheResponse.java
@@ -5,9 +5,9 @@
 
 package org.opensearch.knn.plugin.transport;
 
-import org.opensearch.action.support.DefaultShardOperationFailedException;
+import org.opensearch.core.action.support.DefaultShardOperationFailedException;
 import org.opensearch.action.support.broadcast.BroadcastResponse;
-import org.opensearch.common.io.stream.StreamInput;
+import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.xcontent.ToXContentObject;
 
 import java.io.IOException;